+++ index.js
@@ -20,10 +20,12 @@ const resembleJS = require('./lib/resemble');
  * @param {boolean} options.debug Add some extra logging and always save the image difference (default:false)
  * @param {string} options.formatImageName Custom variables for Image Name (default:{tag}-{browserName}-{width}x{height}-dpr-{dpr})
  * @param {boolean} options.disableCSSAnimation Disable all css animations on a page (default:false)
+ * @param {boolean} options.hideScrollBars Hide all scrolls on a page (default:true)
  * @param {boolean} options.nativeWebScreenshot If a native screenshot of a device (complete screenshot) needs to be taken (default:false)
  * @param {boolean} options.blockOutStatusBar  If the statusbar on mobile / tablet needs to blocked out by default
  * @param {boolean} options.ignoreAntialiasing compare images an discard anti aliasing
  * @param {boolean} options.ignoreColors Even though the images are in colour, the comparison wil compare 2 black/white images
+ * @param {boolean} options.ignoreTransparentPixel Will ignore all pixels that have some transparency in one of the images
  * @param {object} options.androidOffsets Object that will hold custom values for the statusBar, addressBar, addressBarScrolled and toolBar
  * @param {object} options.iosOffsets Object that will hold the custom values for the statusBar, addressBar, addressBarScrolled and toolBar
  *
@@ -63,6 +65,7 @@ class protractorImageComparison {
         this.autoSaveBaseline = options.autoSaveBaseline || false;
         this.debug = options.debug || false;
         this.disableCSSAnimation = options.disableCSSAnimation || false;
+        this.hideScrollBars = options.hideScrollBars !== false;
         this.formatString = options.formatImageName || '{tag}-{browserName}-{width}x{height}-dpr-{dpr}';
 
         this.nativeWebScreenshot = !!options.nativeWebScreenshot;
@@ -70,6 +73,7 @@ class protractorImageComparison {
 
         this.ignoreAntialiasing = options.ignoreAntialiasing || false;
         this.ignoreColors = options.ignoreColors || false;
+        this.ignoreTransparentPixel = options.ignoreTransparentPixel || false;
 
         // OS offsets
         let androidOffsets = options.androidOffsets && typeof options.androidOffsets === 'object' ? options.androidOffsets : {};
@@ -335,6 +339,7 @@ class protractorImageComparison {
      * @param {boolean} compareOptions.blockOutStatusBar blockout the statusbar yes or no, it will override the global
      * @param {boolean} compareOptions.ignoreAntialiasing compare images an discard anti aliasing
      * @param {boolean} compareOptions.ignoreColors Even though the images are in colour, the comparison wil compare 2 black/white images
+     * @param {boolean} compareOptions.ignoreTransparentPixel Will ignore all pixels that have some transparency in one of the images
      * @returns {Promise}
      * @private
      */
@@ -347,6 +352,7 @@ class protractorImageComparison {
         compareOptions.ignoreAntialiasing = 'ignoreAntialiasing' in compareOptions ? compareOptions.ignoreAntialiasing : this.ignoreAntialiasing;
         compareOptions.ignoreColors = 'ignoreColors' in compareOptions ? compareOptions.ignoreColors : this.ignoreColors;
         compareOptions.ignoreRectangles = 'ignoreRectangles' in compareOptions ? compareOptions.ignoreRectangles.push(ignoreRectangles) : ignoreRectangles;
+        compareOptions.ignoreTransparentPixel = 'ignoreTransparentPixel' in compareOptions ? compareOptions.ignoreTransparentPixel : this.ignoreTransparentPixel;
 
         if (this._isMobile() && ((this.nativeWebScreenshot && compareOptions.isScreen) || (this._isIOS())) && blockOutStatusBar) {
             const statusBarHeight = this._isAndroid() ? this.androidOffsets.statusBar : this.iosOffsets.statusBar,
@@ -855,9 +861,9 @@ class protractorImageComparison {
      * @private
      */
     _setCustomTestCSS() {
-        return browser.driver.executeScript(setCSS, this.disableCSSAnimation, this.addressBarShadowPadding, this.toolBarShadowPadding);
+        return browser.driver.executeScript(setCSS, this.disableCSSAnimation, this.hideScrollBars, this.addressBarShadowPadding, this.toolBarShadowPadding);
 
-        function setCSS(disableCSSAnimation, addressBarShadowPadding, toolBarShadowPadding) {
+        function setCSS(disableCSSAnimation, hideScrollBars, addressBarShadowPadding, toolBarShadowPadding) {
             var animation = '* {' +
                     '-webkit-transition-duration: 0s !important;' +
                     'transition-duration: 0s !important;' +
@@ -867,7 +873,7 @@ class protractorImageComparison {
                 scrollBar = '*::-webkit-scrollbar { display:none; !important}',
                 bodyTopPadding = addressBarShadowPadding === 0 ? '' : 'body{padding-top:' + addressBarShadowPadding + 'px !important}',
                 bodyBottomPadding = toolBarShadowPadding === 0 ? '' : 'body{padding-bottom:' + toolBarShadowPadding + 'px !important}',
-                css = disableCSSAnimation ? scrollBar + animation + bodyTopPadding + bodyBottomPadding : scrollBar + bodyTopPadding + bodyBottomPadding,
+                css = (disableCSSAnimation ? animation : '') + (hideScrollBars ? scrollBar : '') + bodyTopPadding + bodyBottomPadding,
                 head = document.head || document.getElementsByTagName('head')[0],
                 style = document.createElement('style');
 
@@ -896,6 +902,8 @@ class protractorImageComparison {
      * browser.protractorImageComparison.checkElement(element(By.id('elementId')), 'imageA', {ignoreAntialiasing: true});
      * // Ignore colors
      * browser.protractorImageComparison.checkElement(element(By.id('elementId')), 'imageA', {ignoreColors: true});
+     * // Ignore alpha pixel
+     * browser.protractorImageComparison.checkElement(element(By.id('elementId')), 'imageA', {ignoreTransparentPixel: true});
      *
      * @param {Promise} element The ElementFinder that is used to get the position
      * @param {string} tag The tag that is used
@@ -904,6 +912,7 @@ class protractorImageComparison {
      * @param {int} options.resizeDimensions the value to increase the size of the element that needs to be saved
      * @param {boolean} options.ignoreAntialiasing compare images an discard anti aliasing
      * @param {boolean} options.ignoreColors Even though the images are in colour, the comparison wil compare 2 black/white images
+     * @param {boolean} options.ignoreTransparentPixel Will ignore all pixels that have some transparency in one of the images
      * @return {Promise} When the promise is resolved it will return the percentage of the difference
      * @public
      */
@@ -936,6 +945,8 @@ class protractorImageComparison {
      * browser.protractorImageComparison.checkFullPageScreen('imageA', {ignoreAntialiasing: true});
      * // Ignore colors
      * browser.protractorImageComparison.checkFullPageScreen('imageA', {ignoreColors: true});
+     * // Ignore alpha pixel
+     * browser.protractorImageComparison.checkFullPageScreen('imageA', {ignoreTransparentPixel: true});
      *
      * @param {string} tag The tag that is used
      * @param {object} options (non-default) options
@@ -973,6 +984,9 @@ class protractorImageComparison {
      * browser.protractorImageComparison.checkScreen('imageA', {ignoreAntialiasing: true});
      * // Ignore colors
      * browser.protractorImageComparison.checkScreen('imageA', {ignoreColors: true});
+     * // Ignore alpha pixel
+     * browser.protractorImageComparison.checkScreen('imageA', {ignoreTransparentPixel: true});
+
      *
      * @param {string} tag The tag that is used
      * @param {object} options (non-default) options
@@ -981,6 +995,7 @@ class protractorImageComparison {
      * @param {boolean} options.disableCSSAnimation enable or disable CSS animation
      * @param {boolean} options.ignoreAntialiasing compare images an discard anti aliasing
      * @param {boolean} options.ignoreColors Even though the images are in colour, the comparison wil compare 2 black/white images
+     * @param {boolean} options.ignoreTransparentPixel Will ignore all pixels that have some transparency in one of the images
      * @return {Promise} When the promise is resolved it will return the percentage of the difference
      * @public
      */